+++ NEWS.md
@@ -1,3 +1,9 @@
+# HealthDataScotland 2.2.0
+
+* Refactor `hopsital` object UI to use have data type and specialty select inputs
+* Refactor `hospital` object summary to have statistics per specialty and data type
+* Remove dark mode
+
 # HealthDataScotland 2.1.1
 
 * Update package disclaimer